requirements.txt
torch>=2.0.0
transformers>=4.37.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.11
accelerate>=0.21.0
bitsandbytes>=0.41.0
einops>=0.7.0

huggingface_hub>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tqdm>=4.65.0
pyyaml>=6.0

pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
scipy>=1.9.0
scikit-learn>=1.2.0

jupyter>=1.0.0